# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/iv-compatibility": {
      "get": {
        "description": "Curated subset of inpatient IV co-administration pairs — high-mortality / high-volume scenarios from King Guide, ASHP, FDA labelling. Includes the documented-deaths cases (ceftriaxone + calcium in neonates), common ICU mistakes (vasopressor + bicarbonate, vancomycin + pip-tazo) and the permissive defaults (norepinephrine Y-site, dexmedetomidine, metronidazole). Filters: q (free text), drug (name or slug), status (compatible|conditional|incompatible|no-data).",
        "operationId": "IvCompatibilityController_list",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "drug",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List curated IV compatibility pairs (Y-site / additive / syringe)",
        "tags": [
          "iv-compatibility"
        ]
      }
    },
    "/api/v1/iv-compatibility/{slug}": {
      "get": {
        "description": "Returns the full pair: compatibility status, applicable routes (y-site / additive / syringe), concentration / diluent / time notes, practical alternative when incompatible, and citations. Reference content only — no patient identity is processed.",
        "operationId": "IvCompatibilityController_get",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get an IV compatibility pair by slug",
        "tags": [
          "iv-compatibility"
        ]
      }
    },
...
    "/api/v1/integrations": {
      "get": {
        "description": "Catalogue of plugin / snippet integrations for OpenMRS, OpenEMR, Bahmni, ERPNext, GNU Health, DHIS2, OpenHIM, CommCare, Epic, Cerner / Oracle Health, Allscripts / Veradigm, athenahealth, plus pure-standard targets (HL7 v2, FHIR R4, CDS Hooks 1.0, SMART on FHIR). Filter by category (open-source / proprietary / standard), method (cds-hooks, smart-on-fhir, fhir-rest, rest, hl7v2, webhook, iframe-embed) or free-text query. Returns summaries; fetch detailed snippets + links by slug.",
        "operationId": "IntegrationsController_list",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List EMR / HMIS integrations VedaMD supports",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/v1/integrations/postman-collection": {
      "get": {
        "description": "Generates a ready-to-import Postman v2.1 collection covering CDS Hooks discovery, every discovered CDS service (POST with realistic sample body), core content endpoints (drugs, conditions, scores, antidotes, ranges, drug-disease, immunization, allergy, notifiable, pgx), the integrations catalogue itself, and a deterministic agentic evaluate example. Bearer auth is pre-configured at the collection level — set the `apiKey` variable to your sandbox or production key after import. Also imports cleanly into Insomnia (via the Postman importer) and Bruno.",
        "operationId": "IntegrationsController_postmanCollection",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Download a Postman v2.1 collection for the API",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/v1/integrations/{slug}/markdown": {
      "get": {
        "description": "Same content as GET /:slug, rendered as Markdown. Paste into Confluence / Notion / GitHub README / internal wiki. Plain text — no PHI, no patient data.",
        "operationId": "IntegrationsController_getMarkdown",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get an EMR / HMIS integration as a Markdown document",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/v1/integrations/{slug}": {
      "get": {
        "description": "Returns description, supported CDS hooks, copy-paste configuration snippets, plugin / documentation links and integrator notes for the requested system.",
        "operationId": "IntegrationsController_get",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get an EMR / HMIS integration detail by slug",
        "tags": [
          "integrations"
        ]
      }
    },
...
    "/api/v1/notifications/channels/providers": {
      "get": {
        "description": "Returns the provider identifiers the platform supports (twilio-sms, twilio-whatsapp, africastalking-sms, smtp, whatsapp-business). Used by the UI channel-create form to populate the provider picker.",
        "operationId": "NotificationsController_providers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List supported notification providers",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/v1/notifications/channels": {
      "get": {
        "operationId": "NotificationsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List configured notification channels for the calling integrator",
        "tags": [
          "notifications"
        ]
      },
      "post": {
        "description": "Submit provider + credentials. Credentials are AEAD-encrypted with the integratorId as additional authenticated data and never returned over the API. A channel must be tested via POST /:id/test before it can dispatch real alerts.",
        "operationId": "NotificationsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a notification channel",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/v1/notifications/channels/{id}": {
      "get": {
        "operationId": "NotificationsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a single channel (summary only — credentials never returned)",
        "tags": [
          "notifications"
        ]
      },
      "patch": {
        "operationId": "NotificationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update channel name / recipient / sender / enabled status, OR rotate credentials",
        "tags": [
          "notifications"
        ]
      },
      "delete": {
        "operationId": "NotificationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a channel",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/v1/notifications/channels/{id}/test": {
      "post": {
        "description": "Uses the channel default recipient unless `to` is provided. Sends a short canned message unless `body` is provided. Returns the provider delivery result.",
        "operationId": "NotificationsController_test",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send a test message through this channel",
        "tags": [
          "notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "VedaMD API",
    "description": "Source-available, stateless Clinical Decision Support API. CDS Hooks + REST/JSON. VedaMD never stores patient data — records remain with your EMR.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}